backend/requirements.txt
# FastAPI
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
sqlalchemy-cockroachdb==2.0.2
alembic==1.14.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Validation
pydantic==2.10.5
pydantic-settings==2.7.1
email-validator==2.2.0

# CORS
python-cors==1.0.0

# AI/LLM
google-generativeai==0.8.3
